Update QAOA.py
Checked the bit order again and made the optimization part faster.

The bits are already "flipped" in the "convert_qubo_to_ising" function so the order is correct now. The first QUBO variable is mapped to the last qubit, the second QUBO variable is mapped to the second last qubit etc.